requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
cssbeautifier==1.15.3
djlint==1.36.4
EditorConfig==0.17.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsbeautifier==1.15.3
json5==0.10.0
libsass==0.23.0
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==2.2.3
pathspec==0.12.1
pyasn1==0.6.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
rsa==4.9
setuptools==75.8.0
six==1.17.0
sympy==1.13.3
tqdm==4.67.1
Werkzeug==3.1.3
